[refactor] Rest Docs 리팩토링 및 release 0.0.1(#31)

+++ src/main/java/upbrella/be/rent/dto/response/StatusDeclarationResponse.java
@@ -10,7 +10,7 @@ import lombok.*;
 public class StatusDeclarationResponse {
 
     private long id;
-    private int umbrellaId;
+    private long umbrellaId;
     private String content;
     private String etc;
 }
